Refactored cluster syncing code.
